server: Complete app file structure re-organization
This extends the re-organization started in the previous commit to parts
that are not not used by the Groups app.
