+++ src/routing/cli.js
@@ -1,0 +1,19 @@
+import * as endpoints from '../endpoints/index.js';
+
+
+export function setup(app) {
+  app.routeAsync("/cli/download/:version/:assetSuffix")
+    .getAsync(endpoints.cli.download);
+
+  app.routeAsync("/cli/download/pr-build/:prId/:assetSuffix")
+    .getAsync(endpoints.cli.downloadPRBuild);
+
+  app.routeAsync("/cli/download/ci-build/:runId/:assetSuffix")
+    .getAsync(endpoints.cli.downloadCIBuild);
+
+  app.routeAsync("/cli/installer/:os")
+    .getAsync(endpoints.cli.installer);
+
+  app.route("/cli")
+    .get((req, res) => res.redirect("https://docs.nextstrain.org/projects/cli/"));
+}
