+++ src/routing/index.js
@@ -1,15 +1,29 @@
 import * as auspice from './auspice.js';
+import * as canary from './canary.js';
 import * as charon from './charon.js';
+import * as cli from './cli.js';
+import * as community from './community.js';
+import * as core from './core.js';
 import * as errors from "./errors.js";
+import * as fetch from './fetch.js';
 import * as groups from './groups.js';
 import * as openid from './openid.js';
+import * as schemas from './schemas.js';
+import * as staging from './staging.js';
 import * as staticSite from './staticSite.js';
 
 export {
   auspice,
+  canary,
   charon,
+  cli,
+  community,
+  core,
   errors,
+  fetch,
   groups,
   openid,
+  schemas,
+  staging,
   staticSite,
 };
